Update lua.json
I am the owner of lua.is-a.dev, and I'm trying my hardest to update my URL for a dreamhost site. I think they use address records, so I updated my script to support that (if I'm correct) and I'm hoping it works. 

These are the nameservers that dreamhost uses:
(This information is already public)
ns1.dreamhost.com 162.159.26.14
ns2.dreamhost.com 162.159.26.81
ns3.dreamhost.com 162.159.27.84

I am using the first one's IP. If this is incorrect please correct me!
Once the script is updated, it should point to their nameservers and theoretically work when connected. I will then make my site from their built-in wordpress editor. The reason is so I can have subpages that instead of just redirecting the user to a bad free wordpress site, is actually owned and operated by me.